[Refactor] #208 - Search 관련 기능 전면 수정
